style tweaks; start of breakpoint observer

+++ src/components/upcoming-date/upcoming-date.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { Box } from 'rebass';
 import { FaRegCalendarCheck } from 'react-icons/fa';
 
-const dateTimeFormat = new Intl.DateTimeFormat('en-US', { month: 'long', day: 'numeric', timeZone: 'UTC' });
+const dateTimeFormat = new Intl.DateTimeFormat('en-US', { month: 'short', day: 'numeric', timeZone: 'UTC' });
 
 const UpcomingDate = ({ date }) => {
   const formatDate = datestr => dateTimeFormat.format(new Date(datestr));
